refactored gameplan/meta info classes into gameplan package.
