+++ src/main/java/paul/TextQuest/gameplan/ItemLibrary.java
@@ -2,10 +2,12 @@
  * @author Paul Dennis
  * Sep 21, 2018
  */
-package paul.TextQuest.entities;
+package paul.TextQuest.gameplan;
 
 import java.util.HashMap;
 import java.util.Map;
+
+import paul.TextQuest.entities.BackpackItem;
 
 public class ItemLibrary {
